requirements.txt
requests
pandas
python-dotenv
flask
flask-cors
schedule

